Cell Successfully Replaced By MonoCell
MonoCell is a Cell, but it is inherited from MonoBehaviour
