+++ Assets/Prefabs/Cell.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 9088512869704228705}
   - component: {fileID: 8789803148466037374}
   - component: {fileID: 6640010112449765558}
+  - component: {fileID: -5053090655174015303}
   m_Layer: 5
   m_Name: Cell
   m_TagString: Untagged
@@ -58,7 +59,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   m_Material: {fileID: 0}
-  m_Color: {r: 1, g: 1, b: 1, a: 0.392}
+  m_Color: {r: 1, g: 0.65552473, b: 0.10849059, a: 0.392}
   m_RaycastTarget: 1
   m_RaycastPadding: {x: 0, y: 0, z: 0, w: 0}
   m_Maskable: 1
@@ -75,3 +76,16 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
+--- !u!114 &-5053090655174015303
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8316105179779688677}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b3cbbc6d9fc017a4cbfcd59c98f4cf31, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  Gravity: 3
