+++ Assets/Scripts/BoardTest.cs
@@ -1,91 +1,91 @@
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-using UnityEngine.UI;
+//using System;
+//using System.Collections;
+//using System.Collections.Generic;
+//using UnityEngine;
+//using UnityEngine.UI;
 
-public class BoardTest : MonoBehaviour
-{
-    private Cell[,] Cells;
-    [SerializeField]
-    private int _width = 5;
-    [SerializeField]
-    private int _height = 5;
-    public Direction Gravity;
+//public class BoardTest : MonoBehaviour
+//{
+//    private Cell[,] Cells;
+//    [SerializeField]
+//    private int _width = 5;
+//    [SerializeField]
+//    private int _height = 5;
+//    public Direction Gravity;
 
-    [SerializeField]
-    private GameObject _board;
+//    [SerializeField]
+//    private GameObject _board;
 
-    public event Action _startCheckingMatch;
+//    public event Action _startCheckingMatch;
 
-    void Start()
-    {
-        _board = gameObject;
-    }
+//    void Start()
+//    {
+//        _board = gameObject;
+//    }
 
-    void InitializeBoard()
-    {
-        Cells = new Cell[_width, _height];
-        for (int i = 0; i < _height; i++)
-        {
-            for (int j = 0; j < _width; j++)
-            {
-                //GameObject cellObject = Instantiate(_cellPrefab, transform);
-                //cellObject.GetComponent<VerticalLayoutGroup>().childControlHeight = true;
-                //Crystal crystal = GenerateCrystalInCell(cellObject);
-                //Cells[j, i] = new Cell(crystal, Gravity, cellObject, null);
-            }
-        }
+//    void InitializeBoard()
+//    {
+//        Cells = new Cell[_width, _height];
+//        for (int i = 0; i < _height; i++)
+//        {
+//            for (int j = 0; j < _width; j++)
+//            {
+//                //GameObject cellObject = Instantiate(_cellPrefab, transform);
+//                //cellObject.GetComponent<VerticalLayoutGroup>().childControlHeight = true;
+//                //Crystal crystal = GenerateCrystalInCell(cellObject);
+//                //Cells[j, i] = new Cell(crystal, Gravity, cellObject, null);
+//            }
+//        }
 
-        for (int i = 0; i < _width; i++)
-        {
-            for (int j = 0; j < _height; j++)
-            {
-                Cells[i, j].SetNeighbors(FillNeighbors(i, j));
-            }
-        }
-    }
+//        for (int i = 0; i < _width; i++)
+//        {
+//            for (int j = 0; j < _height; j++)
+//            {
+//                Cells[i, j].SetNeighbors(FillNeighbors(i, j));
+//            }
+//        }
+//    }
 
-    public void EndSwapping()
-    {
-        StartCheckingMatch();
-    }
-    public void StartCheckingMatch()
-    {
-        Debug.Log("Start Cheking");
-        _startCheckingMatch?.Invoke();
-    }
-    private Neighbors FillNeighbors(int indexI, int indexJ)
-    {
-        Neighbors neighbors = new Neighbors();
-        if (Cells == null)
-            return neighbors;
+//    public void EndSwapping()
+//    {
+//        StartCheckingMatch();
+//    }
+//    public void StartCheckingMatch()
+//    {
+//        Debug.Log("Start Cheking");
+//        _startCheckingMatch?.Invoke();
+//    }
+//    private Neighbors FillNeighbors(int indexI, int indexJ)
+//    {
+//        Neighbors neighbors = new Neighbors();
+//        if (Cells == null)
+//            return neighbors;
 
-        if (indexI > 0)
-            neighbors._left_cell = Cells[indexI - 1, indexJ];
-        if (indexI + 1 < _width)
-            neighbors._right_cell = Cells[indexI + 1, indexJ];
-        if (indexJ > 0)
-            neighbors._top_cell = Cells[indexI, indexJ - 1];
-        if (indexJ + 1 < _height)
-            neighbors._bottom_cell = Cells[indexI, indexJ + 1];
-        return neighbors;
-    }
+//        if (indexI > 0)
+//            neighbors._left_cell = Cells[indexI - 1, indexJ];
+//        if (indexI + 1 < _width)
+//            neighbors._right_cell = Cells[indexI + 1, indexJ];
+//        if (indexJ > 0)
+//            neighbors._top_cell = Cells[indexI, indexJ - 1];
+//        if (indexJ + 1 < _height)
+//            neighbors._bottom_cell = Cells[indexI, indexJ + 1];
+//        return neighbors;
+//    }
 
-    private Crystal GenerateCrystalInCell(GameObject cell)
-    {
-        try
-        {
-            //CrystalData crystalData = _setOfCrystals[UnityEngine.Random.Range(0, _setOfCrystals.Length)];
-            //GameObject crystalPrefab = Instantiate(crystalData.Prefab, cell.transform);
-            //Crystal crystal = crystalPrefab.GetComponent<Crystal>();
-            //crystal.Type = crystalData.Type;
-            return null;
-        }
-        catch (IndexOutOfRangeException)
-        {
-            Debug.LogError("The set of crystals is not filled. Check the field \"Set Of Crystals\"");
-        }
-        return null;
-    }
-}
+//    private Crystal GenerateCrystalInCell(GameObject cell)
+//    {
+//        try
+//        {
+//            //CrystalData crystalData = _setOfCrystals[UnityEngine.Random.Range(0, _setOfCrystals.Length)];
+//            //GameObject crystalPrefab = Instantiate(crystalData.Prefab, cell.transform);
+//            //Crystal crystal = crystalPrefab.GetComponent<Crystal>();
+//            //crystal.Type = crystalData.Type;
+//            return null;
+//        }
+//        catch (IndexOutOfRangeException)
+//        {
+//            Debug.LogError("The set of crystals is not filled. Check the field \"Set Of Crystals\"");
+//        }
+//        return null;
+//    }
+//}
