+++ Assets/Scripts/Direction.cs
@@ -1,0 +1,8 @@
+﻿public enum Direction
+{
+    Left,
+    Right,
+    Top,
+    Bottom,
+    None
+}
