+++ Assets/Scripts/Direction.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4feb42c70ffeac9438b0f2bffbf5deee
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
