+++ Assets/Scripts/Neighbors.cs
@@ -1,0 +1,7 @@
+﻿public struct Neighbors
+{
+    public MonoCell _left_cell;
+    public MonoCell _right_cell;
+    public MonoCell _top_cell;
+    public MonoCell _bottom_cell;
+}
